refactor: Move all the main registration to run the app into one file
* Run deadlock retry in queries too
* Fix stack overflow exception in `SingleQueryDelegator`
