+++ CSharp/src/BusinessApp.WebApi/ServiceRegistrations/MacroScopeWrappingHandler.cs
@@ -3,7 +3,12 @@ namespace BusinessApp.WebApi
     using BusinessApp.App;
     using System.Collections.Generic;
 
-    // [Proxy(typeof(ValidationRequestDecorator<,>))]
+    /// <summary>
+    /// Proxies the request to the <see cref="BatchRequestDelegator{TRequest, TResponse}"/>
+    /// </summary>
+    /// <typeparam name="TRequest"></typeparam>
+    /// <typeparam name="TResponse"></typeparam>
+    /// <remarks>Only purpose is signal in the pipeline that a macro caused this</remarks>
     internal sealed class MacroScopeWrappingHandler<TRequest, TResponse> :
          BatchScopeWrappingHandler<BatchRequestDelegator<TRequest, TResponse>, IEnumerable<TRequest>, IEnumerable<TResponse>>
     {
